requirements.txt
tqdm
time
colorama
faker
random
datetime
os
request
webbrowser
pathlib